+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: wjhoe <wjhoe@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/06 16:35:04 by wjhoe             #+#    #+#             */
-/*   Updated: 2025/05/07 00:07:10 by wjhoe            ###   ########.fr       */
+/*   Updated: 2025/05/15 14:34:06 by wjhoe            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
